+++ en_file_uploader/lib/en_file_uploader.dart
@@ -1,10 +1,12 @@
 /// file uploader package used to handle file upload.
 ///
 /// follow `README.md` and [FileUploadController]
-/// to undestartand how this library can be used.
+/// to understand how this library can be used.
 library en_file_uploader;
 
 import 'package:en_file_uploader/src/controller/file_upload_controller.dart';
+
+export 'package:cross_file/cross_file.dart' show XFile;
 
 export 'src/config.dart';
 export 'src/controller/file_upload_controller.dart';