+++ en_file_uploader/test/src/file_uploader_test.dart
@@ -82,8 +82,6 @@ void main() {
         expect(handler.statusCount, 1);
         expect(handler.total, 6);
       });
-
-      test('should retry everything', () {});
     },
   );
 }